Change the order in which undo and hooks are called
Previously, undo recorded a snapshot before running the hook. This
resulted in the wrong behaviour noticable with node_release_hooks:
adding a coreference arrow was not undoable, but it became part of the
following action and would be undone together with it.

This fixes the behaviour and changes made by node release hooks can be
normally undone.